rename to resource view
